% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiotrackingeu.R
\name{angle_between}
\alias{angle_between}
\title{Calculates the smaller angle between two vectors.}
\usage{
angle_between(angle_a, angle_b)
}
\arguments{
\item{angle_a}{Angle of vector a.}

\item{angle_b}{Angle of vector b.}
}
\value{
The smaller angle between both vectors - always less than or equal to 180 degrees.
}
\description{
Calculates the smaller angle between two vectors.
}
\examples{
angle_between(30,77)
#returns 47
angle_between(20,350)
#returns 30
}
\author{
Dipl.-Phys. Ralf Zeidler, \email{ralf.zeidler@fridata.de}
}
