% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiotrackingeu.R
\name{calc_angle}
\alias{calc_angle}
\title{Calculates the DoA using the signal strengths at two antennas.}
\usage{
calc_angle(sig_a, sig_b, angle_a, angle_b, dbLoss, option)
}
\arguments{
\item{sig_a}{Signal strength at antenna a.}

\item{sig_b}{Signal strength at antenna b.}

\item{angle_a}{Orientation of antenna a.}

\item{angle_b}{Orientation of antenna b.}

\item{dbLoss}{Signal drop at 90 degrees in dB.}

\item{option}{A string specifying the approximation method. "arccos" can only be used with antennas at 90 degree intervals, "linear" is more rough and "automatic" chooses depending on the angle between antennas.}
}
\value{
Approximation of the DoA.
}
\description{
Calculates the DoA using the signal strengths at two antennas.
}
\examples{
calc_angle(20,30,0,90,15,"arccos")
calc_angle(20,20,0,90,15,"arccos")
}
\author{
Dipl.-Phys. Ralf Zeidler, \email{ralf.zeidler@fridata.de}
}
