% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiotrackingeu.R
\name{centroid_fun}
\alias{centroid_fun}
\title{Calculates an average position using a moving average for the x and y coordinates.}
\usage{
centroid_fun(tri_data, time, s_time, method = "mean")
}
\arguments{
\item{tri_data}{Data frame with triangulated positions. Expects columns \code{timestamp,freq_tag, pos.X, pos.utm.X, pos.utm.Y}.}

\item{time}{aggregated time}

\item{s_time}{Time window the average is applied to.}

\item{method}{What average to be used: "mean" or "median".}
}
\value{
Data frame holding the moving average of the positions. Holds columns \code{timestamp, freq_tag, pos.utm.X, pos.utm.Y, utm.zone, X, Y}.
}
\description{
Calculates an average position using a moving average for the x and y coordinates.
}
\author{
Dipl.-Phys. Ralf Zeidler, \email{ralf.zeidler@fridata.de}
}
