% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiotrackingeu.R
\name{doa}
\alias{doa}
\title{Wrapper function for doa_internal - requires time matched signals}
\usage{
doa(signals, receivers, dBLoss = 14, live_mode = FALSE,
  live_update_interval = 15, progress = FALSE)
}
\arguments{
\item{signals}{Data frame with certain column names containing the signal logs.}

\item{receivers}{Data frame with certain column names containing receiver / antenna information (location, orientation).}

\item{dBLoss}{Signal drop at 90 degrees in dB - passed to calc_angle().}

\item{live_mode}{Information required by the rteu Logger shiny app.}

\item{live_update_interval}{Information required by the rteu Logger shiny app.}

\item{progress}{Whether this call is wrapped in a \code{shiny::withProgress} call.}
}
\value{
Data frame containing the DoAs.
}
\description{
Wrapper function for doa_internal - requires time matched signals
}
\examples{
logger_data_tm <- time_match_signals(ld_small)
doa(logger_data_tm,receiver_data)

}
\author{
Dipl.-Phys. Ralf Zeidler, \email{ralf.zeidler@fridata.de}
}
