% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiotrackingeu.R
\name{doa_internal}
\alias{doa_internal}
\title{Chooses which antennas to use to calculate the DoA.}
\usage{
doa_internal(data, time_to_look_for, dBLoss = 14,
  doa_approx = "automatic", progress = FALSE)
}
\arguments{
\item{data}{Data frame with certain column names containing signal logs and receiver data.}

\item{time_to_look_for}{Vector of timestamps to be analyzed.}

\item{dBLoss}{Signal drop at 90 degrees in dB - passed to calc_angle().}

\item{doa_approx}{A string specifying DoA approximation method used in calc_angle(). One of "arccos", "linear", "automatic".}

\item{progress}{Whether this call is wrapped in a \code{shiny::withProgress} call.}
}
\value{
dataframe containing the DoAs.
}
\description{
Chooses which antennas to use to calculate the DoA.
}
\author{
Dipl.-Phys. Ralf Zeidler, \email{ralf.zeidler@fridata.de}
}
