% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiotrackingeu.R
\docType{data}
\name{ld_small}
\alias{ld_small}
\title{Data set containing signal recordings of one bat on two stations.}
\description{
It is a subst of logger_data to speed up process time while running the examples.
}
\references{
\url{www.radio-tracking.eu}
}
\author{
Dipl.-Phys. Ralf Zeidler, \email{ralf.zeidler@fridata.de}

Ralf Zeidler \email{ralf.zeidler@fridata.de}
}
