% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiotrackingeu.R
\docType{data}
\name{logger_data}
\alias{logger_data}
\title{Data set containing signal recordings of one bat on two stations.}
\description{
Data set containing signal recordings of one bat on two stations.
}
\references{
\url{www.radio-tracking.eu}
}
\author{
Dipl.-Phys. Ralf Zeidler, \email{ralf.zeidler@fridata.de}

Ralf Zeidler \email{ralf.zeidler@fridata.de}
}
