% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiotrackingeu.R
\docType{data}
\name{receiver_data}
\alias{receiver_data}
\title{Data set containing position and orientation of two tracking stations.}
\description{
Data set containing position and orientation of two tracking stations.
}
\references{
\url{www.radio-tracking.eu}
}
\author{
Dipl.-Phys. Ralf Zeidler, \email{ralf.zeidler@fridata.de}

Ralf Zeidler \email{ralf.zeidler@fridata.de}
}
