% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiotrackingeu.R
\name{smooth_to_time_match}
\alias{smooth_to_time_match}
\title{Matches timestamps by fitting a smoothing spline.}
\usage{
smooth_to_time_match(data, spar_value = 0.01, progress = FALSE)
}
\arguments{
\item{data}{Data frame with certain column names containing the signal logs.}

\item{spar_value}{Smoothing coefficient to be used with stats::smooth.spline.}

\item{progress}{Whether this call is wrapped in a \code{shiny::withProgress} call.}
}
\value{
Data frame with matched timestamps.
}
\description{
Matches timestamps by fitting a smoothing spline.
}
\examples{
smooth_to_time_match(ld_small,spar_value=0.01, progress=FALSE)
}
\author{
Dipl.-Phys. Ralf Zeidler, \email{ralf.zeidler@fridata.de}
}
