% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiotrackingeu.R
\name{smooth_to_time_match_bearings}
\alias{smooth_to_time_match_bearings}
\title{uses a spline to time match the signal strength between several stations}
\usage{
smooth_to_time_match_bearings(data, receivers, spar_value = 0.01,
  progress = FALSE)
}
\arguments{
\item{data}{Data frame containing logger data.}

\item{receivers}{Data frame containing receiver information.}

\item{spar_value}{Smoothing factor of the spline.}

\item{progress}{Whether this call is wrapped in a \code{shiny::withProgress} call.}
}
\value{
return dataframe containg time matched signal strength using spline
}
\description{
uses a spline to time match the signal strength between several stations
}
\author{
Dipl.-Phys. Ralf Zeidler, \email{ralf.zeidler@fridata.de}
}
