% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiotrackingeu.R
\name{time_match_signals}
\alias{time_match_signals}
\title{Matches timestamps by dividing into time slots.}
\usage{
time_match_signals(data, station_time_error = 0.3, progress = FALSE)
}
\arguments{
\item{data}{Data frame with certain column names containing the signal logs.}

\item{station_time_error}{Maximum length of the time slots.}

\item{progress}{Whether this call is wrapped in a \code{shiny::withProgress} call.}
}
\value{
Data frame with matched timestamps.
}
\description{
Matches timestamps by dividing into time slots.
}
\examples{
time_match_signals(ld_small)
}
\author{
Dipl.-Phys. Ralf Zeidler, \email{ralf.zeidler@fridata.de}
}
