% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiotrackingeu.R
\name{timematch_inter}
\alias{timematch_inter}
\title{function to match times between two or more station}
\usage{
timematch_inter(data, inter_error = 0.6)
}
\arguments{
\item{data}{bearing data}

\item{inter_error}{maximal error between stations}
}
\value{
data frame with matched timestamps
}
\description{
function to match times between two or more station
}
\examples{
#registerDoParallel(cores=2)
logger_data_tm <- time_match_signals(ld_small)
doa <- doa(logger_data_tm,receiver_data)
timematch_inter(doa)
#stopImplicitCluster()

}
\author{
Dipl.-Phys. Ralf Zeidler, \email{ralf.zeidler@fridata.de}
}
