% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiotrackingeu.R
\name{tri_centroid}
\alias{tri_centroid}
\title{Calculates the centroid of intersection points of the given bearings and checks if intersection angles are in allowed range.}
\usage{
tri_centroid(tmp_fts, angles_allowed)
}
\arguments{
\item{tmp_fts}{Data frame of bearings. Expects columns \code{utm.X, utm.Y, utm.zone, angle}.}

\item{angles_allowed}{Vector of lower and upper limits for angles between bearings.}
}
\value{
Data frame holding centroid position as lng/lat \code{X,Y} and UTM \code{utm.X, utm.Y}. Will hold \code{NA} if triangulation yields no result.
}
\description{
Calculates the centroid of intersection points of the given bearings and checks if intersection angles are in allowed range.
}
\author{
Dipl.-Phys. Ralf Zeidler, \email{ralf.zeidler@fridata.de}
}
