% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiotrackingeu.R
\name{tri_one}
\alias{tri_one}
\title{Approximates a position using one bearing and an estimation of the distance.}
\usage{
tri_one(tmp_fts)
}
\arguments{
\item{tmp_fts}{Data frame of bearings. Expects columns \code{utm.X, utm.Y, utm.zone, angle, strength}.}
}
\value{
Data frame holding approximated position as lng/lat \code{X,Y} and UTM \code{utm.X, utm.Y}.
}
\description{
Approximates a position using one bearing and an estimation of the distance.
}
\author{
Dipl.-Phys. Ralf Zeidler, \email{ralf.zeidler@fridata.de}
}
