% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiotrackingeu.R
\name{tri_two}
\alias{tri_two}
\title{Calculates the position using only the bearings of two the strongest signals.}
\usage{
tri_two(tmp_fts, angles_allowed)
}
\arguments{
\item{tmp_fts}{Data frame of bearings. Expects columns \code{utm.X, utm.Y, utm.zone, angle}.}

\item{angles_allowed}{Vector of lower and upper limits for angles between bearings.}
}
\value{
Data frame holding the calculated position as lng/lat \code{X,Y} and UTM \code{utm.X, utm.Y}. Will hold \code{NA} if triangulation yields no result.
}
\description{
Calculates the position using only the bearings of two the strongest signals.
}
\author{
Dipl.-Phys. Ralf Zeidler, \email{ralf.zeidler@fridata.de}
}
