% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiotrackingeu.R
\name{triang}
\alias{triang}
\title{Does the actual triangulation. Assumes GK coordinates or UTM coordinates within the same zone.}
\usage{
triang(x1, y1, alpha1, x2, y2, alpha2)
}
\arguments{
\item{x1}{Easting of the first station.}

\item{y1}{Northing of the first station.}

\item{alpha1}{DoA of the first station.}

\item{x2}{Easting of the second station.}

\item{y2}{Northing of the second station.}

\item{alpha2}{DoA of the second station.}
}
\value{
Vector with Easting \code{x} and Northing \code{y} of triangulated position.
}
\description{
Does the actual triangulation. Assumes GK coordinates or UTM coordinates within the same zone.
}
\examples{
triang(0,0,45,10,0,315)

}
\author{
Dipl.-Phys. Ralf Zeidler, \email{ralf.zeidler@fridata.de}
}
