% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiotrackingeu.R
\name{triangulate}
\alias{triangulate}
\title{Calculates the intersection of bearings of the same frequency and time interval.}
\usage{
triangulate(receivers, bearings, only_one = FALSE,
  time_error_inter_station = 0.6, angles_allowed = c(30, 150),
  tri_option = "centroid", tm_method = "spline", spar = 0.01,
  progress = FALSE)
}
\arguments{
\item{receivers}{Data frame of all receivers with station names and positions.}

\item{bearings}{Data frame produced by doa().}

\item{only_one}{Whether a location should be estimated based on signal strength if only one bearing is available in a time slot.}

\item{time_error_inter_station}{Accepted time discrepancy between station clocks in seconds. Used in timematch_inter()}

\item{angles_allowed}{Vector of lower and upper limits for angles between bearings.}

\item{tri_option}{What to do if more than two bearings per time slot are available: "two_strongest" - just consider the two bearings with the strongest signal, "centroid" - calculate a centroid of all intersection points}

\item{tm_method}{Which method of time matching to use: "tm" - time slots, "spline" - spline smoothing See \code{\link{smooth_to_time_match}} and \code{\link{time_match_signals}}.}

\item{spar}{Smoothing factor for the spline to be used with \code{tm_method="spline"}.}

\item{progress}{Whether this call is wrapped in a \code{shiny::withProgress} call.}
}
\value{
Data frame containing triangulated UTM coordinates.
}
\description{
Calculates the intersection of bearings of the same frequency and time interval.
}
\examples{
logger_data_tm <- time_match_signals(ld_small)
#registerDoParallel(cores=2)
doa <- doa(logger_data_tm,receiver_data)
triangulate(receiver_data,doa)
#stopImplicitCluster()

}
\author{
Dipl.-Phys. Ralf Zeidler, \email{ralf.zeidler@fridata.de}
}
