% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiotrackingeu.R
\name{utmtowgs}
\alias{utmtowgs}
\title{Converts UTM coordinates to longitude/latitude based on WGS84 Datum.}
\usage{
utmtowgs(x, y, zone)
}
\arguments{
\item{x}{Easting}

\item{y}{Northing}

\item{zone}{Zone}
}
\value{
Data frame holding longitude (X) and latitude (Y) columns.
}
\description{
Converts UTM coordinates to longitude/latitude based on WGS84 Datum.
}
\author{
Dipl.-Phys. Ralf Zeidler, \email{ralf.zeidler@fridata.de}
}
