% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiotrackingeu.R
\name{wgstoutm}
\alias{wgstoutm}
\title{Converts longitude/latitude coordinates to UTM based WGS84 Datum.}
\usage{
wgstoutm(x, y)
}
\arguments{
\item{x}{Longitude}

\item{y}{Latitude}
}
\value{
UTM coordinates in a data frame with columns X, Y and zone.
}
\description{
Converts longitude/latitude coordinates to UTM based WGS84 Datum.
}
\author{
Dipl.-Phys. Ralf Zeidler, \email{ralf.zeidler@fridata.de}
}
